requirements.txt
# Core
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
pydantic>=2.10.0
pydantic-settings>=2.7.0
python-dotenv>=1.0.0

# Database
sqlalchemy[asyncio]>=2.0.36
asyncpg>=0.30.0
alembic>=1.14.0
pgvector>=0.3.6

# AI / Agents
langgraph>=0.2.60
langchain-core>=0.3.28
langchain-anthropic>=0.3.4
anthropic>=0.42.0
litellm>=1.55.0

# Scraping & Data
firecrawl-py>=1.8.0
httpx>=0.28.0
beautifulsoup4>=4.12.0

# Email
resend>=2.5.0

# Phone (Retell AI)
retell-sdk>=4.0.0

# Gmail API (inbox monitoring)
google-auth>=2.27.0
google-auth-oauthlib>=1.2.0
google-api-python-client>=2.114.0

# Task Queue
celery[redis]>=5.4.0
redis>=5.2.0

# Observability
langfuse>=2.57.0

# Utilities
python-multipart>=0.0.20
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
tenacity>=9.0.0
numpy>=2.0.0

# Testing
pytest>=8.3.0
pytest-asyncio>=0.25.0
pytest-httpx>=0.35.0
